firmware/test/requirements.txt
pytest
pytest_embedded
pyserial
